+++ crates/cli/src/enums.rs
@@ -1,31 +1,35 @@
 use clap::ArgEnum;
+use serde::{Deserialize, Serialize};
 use strum_macros::Display;
 
-#[derive(ArgEnum, Clone, Debug, Display)]
+#[derive(ArgEnum, Clone, Debug, Default, Display)]
 pub enum CacheMode {
     Off,
     Read,
+    #[default]
     Write,
 }
 
-impl Default for CacheMode {
-    fn default() -> Self {
-        CacheMode::Write
-    }
-}
-
-#[derive(ArgEnum, Clone, Debug, Display)]
+#[derive(ArgEnum, Clone, Debug, Default, Display)]
 pub enum LogLevel {
     Off,
     Error,
     Warn,
+    #[default]
     Info,
     Debug,
     Trace,
 }
 
-impl Default for LogLevel {
-    fn default() -> Self {
-        LogLevel::Info
-    }
+#[derive(ArgEnum, Clone, Copy, Debug, Deserialize, Display, Default, PartialEq, Serialize)]
+#[serde(rename_all = "lowercase")]
+pub enum TouchedStatus {
+    Added,
+    #[default]
+    All,
+    Deleted,
+    Modified,
+    Staged,
+    Unstaged,
+    Untracked,
 }