+++ crates/config/src/project/task.rs
@@ -37,31 +37,21 @@ fn validate_outputs(list: &[String]) -> Result<(), ValidationError> {
     Ok(())
 }
 
-#[derive(Clone, Debug, Deserialize, JsonSchema, PartialEq, Serialize)]
+#[derive(Clone, Debug, Default, Deserialize, JsonSchema, PartialEq, Serialize)]
 #[serde(rename_all = "lowercase")]
 pub enum TaskType {
+    #[default]
     Node,
     System,
 }
 
-impl Default for TaskType {
-    fn default() -> Self {
-        TaskType::Node
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, JsonSchema, PartialEq, Serialize)]
+#[derive(Clone, Debug, Default, Deserialize, JsonSchema, PartialEq, Serialize)]
 #[serde(rename_all = "lowercase")]
 pub enum TaskMergeStrategy {
+    #[default]
     Append,
     Prepend,
     Replace,
-}
-
-impl Default for TaskMergeStrategy {
-    fn default() -> Self {
-        TaskMergeStrategy::Append
-    }
 }
 
 #[derive(Clone, Debug, Deserialize, JsonSchema, PartialEq, Serialize, Validate)]