+++ crates/toolchain/src/pms/yarn.rs
@@ -26,7 +26,7 @@ impl YarnTool {
         let install_dir = node.get_install_dir()?.clone();
 
         Ok(YarnTool {
-            bin_path: install_dir.join(node::get_bin_name_suffix("yarn", "cmd", false)),
+            bin_path: node::find_package_manager_bin(&install_dir, "yarn"),
             config: config.to_owned(),
             install_dir,
             log_target: String::from("moon:toolchain:yarn"),
@@ -163,10 +163,7 @@ impl Installable<NodeTool> for YarnTool {
 impl Executable<NodeTool> for YarnTool {
     async fn find_bin_path(&mut self, node: &NodeTool) -> Result<(), ToolchainError> {
         // If the global has moved, be sure to reference it
-        let bin_path = node
-            .get_npm()
-            .get_global_dir()?
-            .join(node::get_bin_name_suffix("yarn", "cmd", false));
+        let bin_path = node::find_package_manager_bin(node.get_npm().get_global_dir()?, "yarn");
 
         if bin_path.exists() {
             self.bin_path = bin_path;